refactor(media-providers): extract media kind constants from nav component
