+++ src/app/(dashboard)/dashboard/media-providers/components/mediaKinds.ts
@@ -1,0 +1,22 @@
+export type MediaKind =
+  | "embedding"
+  | "image"
+  | "imageToText"
+  | "tts"
+  | "stt"
+  | "webSearch"
+  | "webFetch"
+  | "video"
+  | "music";
+
+export const MEDIA_KINDS: MediaKind[] = [
+  "embedding",
+  "image",
+  "imageToText",
+  "tts",
+  "stt",
+  "webSearch",
+  "webFetch",
+  "video",
+  "music",
+];
